Moved NWItem and NWStatus into the project file as the classes are no longer called from elsewhere
