+++ nw/project/__init__.py
@@ -3,15 +3,11 @@
 from nw.project.backup import NWBackup
 from nw.project.document import NWDoc
 from nw.project.index import NWIndex
-from nw.project.item import NWItem
 from nw.project.project import NWProject
-from nw.project.status import NWStatus
 
 __all__ = [
     "NWBackup",
     "NWDoc",
     "NWIndex",
-    "NWItem",
     "NWProject",
-    "NWStatus",
 ]
