+++ tests/test_item.py
@@ -9,8 +9,7 @@ from lxml import etree
 from nwdummy import DummyMain
 
 from nw.config import Config
-from nw.project.project import NWProject
-from nw.project.item import NWItem
+from nw.project.project import NWProject, NWItem
 from nw.constants import nwItemClass, nwItemType, nwItemLayout
 
 theConf = Config()
